src/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap4==2024.1
Django==5.1.1
django-bootstrap4==24.4
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
django-localflavor==4.0
environ==1.0
package-name==0.1
pillow==10.4.0
python-dotenv==1.0.1
python-stdnum==1.20
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1

